Fix clang++ errors.
I might add clang++ as a compiler to the CI workflow later, so that I
can verify that quadelect can be compiled with other compilers than
just g++. That will require some rewriting of the CMakeLists file so
that clang++ doesn't get confused by g++-only parameters, though.

+++ src/random/random.cc
@@ -75,7 +75,7 @@ long double rng::ldrand() {
 
 	const uint64_t maxVal = ~((uint64_t)0); // maximum value we'll get.
 
-	const long double RS_SCALE = (1.0 / (1.0 + maxVal));
+	const long double RS_SCALE = (1.0 / (1.0 + (long double)maxVal));
 
 	long double d;
 	do {
